Add pomodoro notifications.

+++ app/src/main/java/com/sarangem/zenwell/ui/focusscreen/PomodoroTimerDisplay.kt
@@ -21,7 +21,7 @@ import com.sarangem.zenwell.ui.theme.Orbitron
 
 @Composable
 fun PomodoroTimerDisplay(
-    elapsedTime: Long,
+    elapsedTime: Int,
     isWorkTime: Boolean
 ) {
     Column(
